frontend/requirements.txt
flask
pyjwt
msal
python-dotenv